feat(maven): initial commit, maven ready, sample pom, name changes
