+++ .gitignore
@@ -9,3 +9,9 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+settings.xml
+target/
+.settings/
+.project
+.classpath
+pom.xml
